refactor: organise lora JS logic into folders
