+++ nodes/web/lora.js
@@ -1,117 +1,5 @@
 import { app } from "../../../scripts/app.js";
-
-function handleNumSlots(node, widget) {
-  const numSlots = parseInt(widget.value);
-
-  if (!node.widgetsHidden) {
-    node.widgetsHidden = [];
-  }
-
-  for (let i = 1; i <= 10; i++) {
-    const widgets = [...node.widgets, ...node.widgetsHidden].filter((w) =>
-      w.name.endsWith(`_${i}`),
-    );
-
-    widgets.forEach((w) => {
-      if (i <= numSlots) {
-        const hiddenIndex = node.widgetsHidden.indexOf(w);
-        const widgetIndex = node.widgets.indexOf(w);
-        if (hiddenIndex !== -1 && widgetIndex === -1) {
-          node.widgets.push(w);
-          node.widgetsHidden.splice(hiddenIndex, 1);
-        }
-      } else {
-        const widgetIndex = node.widgets.indexOf(w);
-        if (widgetIndex !== -1) {
-          node.widgetsHidden.push(w);
-          node.widgets.splice(widgetIndex, 1);
-        }
-      }
-    });
-  }
-
-  handleMode(node, undefined);
-}
-
-function handleMode(node, widget) {
-  const modeWidget = node.widgets.find((w) => w.name === "mode");
-  if (!modeWidget) return;
-
-  if (!node.widgetsHiddenModes) {
-    node.widgetsHiddenModes = [];
-  }
-
-  const isSimple = modeWidget.value === "Simple";
-  const allWidgets = [...node.widgets, ...node.widgetsHiddenModes];
-
-  allWidgets.forEach((w) => {
-    const isWeightWidget = w.name.startsWith("weight_");
-    const isAdvancedWidget =
-      w.name.startsWith("model_weight_") || w.name.startsWith("clip_weight_");
-
-    if (isSimple) {
-      if (isWeightWidget) {
-        const slotNum = w.name.replace("weight_", "");
-        const loraNameWidget = node.widgets.findIndex(
-          (widget) => widget.name === `lora_name_${slotNum}`,
-        );
-
-        if (loraNameWidget !== -1) {
-          const hiddenIndex = node.widgetsHiddenModes.indexOf(w);
-          const widgetIndex = node.widgets.indexOf(w);
-          if (hiddenIndex !== -1 && widgetIndex === -1) {
-            node.widgets.splice(loraNameWidget + 1, 0, w);
-            node.widgetsHiddenModes.splice(hiddenIndex, 1);
-          }
-        }
-      }
-      if (isAdvancedWidget) {
-        const widgetIndex = node.widgets.indexOf(w);
-        if (widgetIndex !== -1) {
-          node.widgetsHiddenModes.push(w);
-          node.widgets.splice(widgetIndex, 1);
-        }
-      }
-    } else {
-      if (isWeightWidget) {
-        const widgetIndex = node.widgets.indexOf(w);
-        if (widgetIndex !== -1) {
-          node.widgetsHiddenModes.push(w);
-          node.widgets.splice(widgetIndex, 1);
-        }
-      }
-      if (isAdvancedWidget) {
-        const slotNum = w.name.replace("model_weight_", "").replace("clip_weight_", "");
-        const loraNameWidget = node.widgets.findIndex(
-          (widget) => widget.name === `lora_name_${slotNum}`,
-        );
-
-        if (loraNameWidget !== -1) {
-          const hiddenIndex = node.widgetsHiddenModes.indexOf(w);
-          const widgetIndex = node.widgets.indexOf(w);
-          if (hiddenIndex !== -1 && widgetIndex === -1) {
-            node.widgets.splice(loraNameWidget + 1, 0, w);
-            node.widgetsHiddenModes.splice(hiddenIndex, 1);
-          }
-        }
-      }
-    }
-  });
-
-  node.setSize(node.computeSize());
-}
-
-const nodeWidgetHandlers = {
-  num_slots: handleNumSlots,
-  mode: handleMode,
-};
-
-function widgetLogic(node, widget, isInitial = false) {
-  const handler = nodeWidgetHandlers[widget.name];
-  if (handler && !isInitial) {
-    handler(node, widget);
-  }
-}
+import { handleMode, handleNumSlots, widgetLogic } from "./helpers/lora/main.js";
 
 const ext = {
   name: "signature.lora_stacker",
@@ -125,9 +13,7 @@ const ext = {
         Object.defineProperty(w, "value", {
           get() {
             let valueToReturn =
-              originalDescriptor && originalDescriptor.get
-                ? originalDescriptor.get.call(w)
-                : widgetValue;
+              originalDescriptor && originalDescriptor.get ? originalDescriptor.get.call(w) : widgetValue;
             return valueToReturn;
           },
           set(newVal) {
